Examples folder created & examples rewritten!
